# Optimization: convert point slice indices to floats