@andy-sweet 2022-06-09
> What happens if you change the type from `float` to `np.data.dtype` (or we know that it will be `float`)?

Do you mean `self.data.dtype`? I think that can be anything (e.g. `int` or `float`) as we don't force the dtype in [`fix_data_points`](https://github.com/napari/napari/blob/1546f18ecc13364d5415623a9c11ed760ff043e2/napari/layers/points/_points_utils.py#L210), though if you manually add points I think `Points.data` will be upcast to `float`.

I think the elements in `dims_indices` that are not `slice` objects are currently guaranteed to be `float` because `Dims.point` contains floats and `Layer._slice_indices` only casts the transformed point to `int` for all layers except `Points`.

Therefore, I think using `dtype=self.data.dtype` will generally downcast `dims_indices`, which will be lossy in general, and should be avoided.

Also, casting between int and float when doing operations is not too costly. But casting between int/float and object is a disaster likely for a number of reasons including memory compactness.

```python
In [55]: n = 10_000_000

In [56]: data_float = 500 * np.random.rand(n)

In [57]: data_int = np.random.randint(500, size=n)

In [58]: %timeit data_float - np.array([250], dtype=float)
18.7 ms ± 218 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

In [59]: %timeit data_float - np.array([250], dtype=int)
18.5 ms ± 645 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

In [60]: %timeit data_float - np.array([250], dtype=object)
531 ms ± 14.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [61]: %timeit data_int - np.array([250], dtype=float)
21.5 ms ± 495 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [62]: %timeit data_int - np.array([250], dtype=int)
16.2 ms ± 436 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

In [63]: %timeit data_int - np.array([250], dtype=object)
549 ms ± 7.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```